Parse GLTF binary files with PNG textures (#2)
The PR adds the ability to parse GLB files containing textures and produce a mesh.

GLB files are the binary representation of GLTF files. They contain two main sections, a json structure describing the scene, meshes, materials, lights etc. and a binary section containing the geometry and the textures. Per the spec, the binary section can be a separate file/s but the parser does not support that.

Since the GLB is mostly a container format, two additional parsers were needed - JSON and PNG. The JSON parser takes in a buffer and generates a tree structure. I've described it in more detail here: https://marsp0.github.io/articles/simple-json-parser.html

The PNG parser takes in a buffer and generates a 2d texture from it. The parser includes a DEFLATE stream decoder written from scratch. The parser is single threaded and due to the nature of the format I don't think I'll be able to make it multi-threaded. Having said that, the code that sits on top (the glb parser) could be updated to parse multiple textures at once (added to the todo).